CallGraph.worker. new module, Components.GraphView
