+++ daemon/app/Main.hs
@@ -67,7 +67,6 @@ import GHCSpecter.UI.Types.Event
   ( BackgroundEvent (RefreshUI),
     Event (BkgEv),
   )
-import GHCSpecter.Worker.CallGraph qualified as CallGraph (test)
 import GHCSpecter.Worker.Hie (hieWorker)
 import GHCSpecter.Worker.ModuleGraph (moduleGraphWorker)
 import Options.Applicative qualified as OA
@@ -128,7 +127,7 @@ listener socketFile var = do
           CMSession s' -> do
             let mgi = sessionModuleGraph s'
             void $ forkIO (moduleGraphWorker var mgi)
-          CMHsSource _modu (HsSourceInfo hiefile) -> do
+          CMHsSource _modu (HsSourceInfo hiefile) ->
             void $ forkIO (hieWorker var hiefile)
           _ -> pure ()
         atomically . modifyTVar' var . updateInbox $ CMBox o
@@ -232,8 +231,11 @@ main = do
           serverSessionRef <- atomically $ newTVar ss
           webServer serverSessionRef
     Temp sessionFile -> do
-      lbs <- BL.readFile sessionFile
+      pure ()
+
+{-      lbs <- BL.readFile sessionFile
       case eitherDecode' lbs of
         Left err -> print err
         Right ss -> do
           CallGraph.test ss
+-}
